+++ packages/core/contracts/test/Memory05.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.5.16;
+
+contract Memory05 {
+  mapping(string => address) a;
+  mapping(bytes => address) b;
+}