+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Support retype from interface to address. ([#687](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/687))
+
 ## 1.20.4 (2022-11-03)
 
 - Support multiple contracts with same name. ([#263](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/263))